# Reduce number of DB queries called for fulfillment logic