--- a/saleor/plugins/admin_email/plugin.py
+++ b/saleor/plugins/admin_email/plugin.py
@@ -1,12 +1,12 @@
 import logging
 from copy import deepcopy
 from dataclasses import asdict
-from typing import Union
+from typing import Callable, Union
 
 from django.conf import settings
 from promise.promise import Promise
 
-from ...core.notify_events import AdminNotifyEvent, NotifyEventType
+from ...core.notify import AdminNotifyEvent, NotifyEventType
 from ...graphql.plugins.dataloaders import EmailTemplatesByPluginConfigurationLoader
 from ..base_plugin import BasePlugin, ConfigurationTypeField, PluginConfigurationType
 from ..email_common import (
@@ -214,7 +214,12 @@ class AdminEmailPlugin(BasePlugin):
             .then(map_templates_to_configuration)
         )
 
-    def notify(self, event: Union[NotifyEventType, str], payload: dict, previous_value):
+    def notify(
+        self,
+        event: Union[NotifyEventType, str],
+        payload_func: Callable[[], dict],
+        previous_value,
+    ):
         if not self.active:
             return previous_value
 
@@ -228,7 +233,7 @@ class AdminEmailPlugin(BasePlugin):
 
         event_func = event_map[event]
         config = asdict(self.config)
-        event_func(payload, config, self)
+        event_func(payload_func, config, self)
 
     @classmethod
     def validate_plugin_configuration(